sp_statement_store::runtime_api
Module statement_store

Host interface

Structs

HostFunctions

Provides implementations for the extern host functions.

Functions

broadcasts

Return the data of all known statements which include all topics and have no DecryptionKey field.

posted

Return the data of all known statements whose decryption key is identified as dest (this will generally be the public key or a hash thereof for symmetric ciphers, or a hash of the private key for symmetric ciphers).

posted_clear

Return the decrypted data of all known statements whose decryption key is identified as dest. The key must be available to the client.

remove

Remove a statement from the store by hash.

statements

Return all statements.

submit_statement

Submit a new new statement. The statement will be broadcast to the network. This is meant to be used by the offchain worker.