sp_statement_store::runtime_api::statement_store

Function posted

pub fn posted(
    match_all_topics: <PassFatPointerAndDecodeSlice<&[Topic]> as RIType>::Inner,
    dest: <PassPointerAndReadCopy<[u8; 32], 32> as RIType>::Inner,
) -> <AllocateAndReturnByCodec<Vec<Vec<u8>>> as RIType>::Inner

Return the data of all known statements whose decryption key is identified as dest (this will generally be the public key or a hash thereof for symmetric ciphers, or a hash of the private key for symmetric ciphers).