Trait sp_statement_store::StatementStore

pub trait StatementStore: Send + Sync {
    // Required methods
    fn statements(&self) -> Result<Vec<(Hash, Statement)>>;
    fn statement(&self, hash: &Hash) -> Result<Option<Statement>>;
    fn broadcasts(&self, match_all_topics: &[Topic]) -> Result<Vec<Vec<u8>>>;
    fn posted(
        &self,
        match_all_topics: &[Topic],
        dest: [u8; 32]
    ) -> Result<Vec<Vec<u8>>>;
    fn posted_clear(
        &self,
        match_all_topics: &[Topic],
        dest: [u8; 32]
    ) -> Result<Vec<Vec<u8>>>;
    fn submit(
        &self,
        statement: Statement,
        source: StatementSource
    ) -> SubmitResult;
    fn remove(&self, hash: &Hash) -> Result<()>;
}

Statement store API.

Required Methods

fn statements(&self) -> Result<Vec<(Hash, Statement)>>

Return all statements.

fn statement(&self, hash: &Hash) -> Result<Option<Statement>>

Get statement by hash.

fn broadcasts(&self, match_all_topics: &[Topic]) -> Result<Vec<Vec<u8>>>

Return the data of all known statements which include all topics and have no DecryptionKey field.

fn posted( &self, match_all_topics: &[Topic], dest: [u8; 32] ) -> Result<Vec<Vec<u8>>>

Return the data of all known statements whose decryption key is identified as dest (this will generally be the public key or a hash thereof for symmetric ciphers, or a hash of the private key for symmetric ciphers).

fn posted_clear( &self, match_all_topics: &[Topic], dest: [u8; 32] ) -> Result<Vec<Vec<u8>>>

Return the decrypted data of all known statements whose decryption key is identified as dest. The key must be available to the client.

fn submit(&self, statement: Statement, source: StatementSource) -> SubmitResult

Submit a statement.

fn remove(&self, hash: &Hash) -> Result<()>

Remove a statement from the store.