sp_statement_store

Trait StatementStore

pub trait StatementStore: Send + Sync {

    // Required methods
    fn statements(&self) -> Result<Vec<(Hash, Statement)>>;
    fn take_recent_statements(&self) -> Result<Vec<(Hash, Statement)>>;
    fn statement(&self, hash: &Hash) -> Result<Option<Statement>>;
    fn has_statement(&self, hash: &Hash) -> bool;
    fn statement_hashes(&self) -> Vec<Hash> ;
    fn statements_by_hashes(
        &self,
        hashes: &[Hash],
        filter: &mut dyn FnMut(&Hash, &[u8], &Statement) -> FilterDecision,
    ) -> Result<(Vec<(Hash, Statement)>, usize)>;
    fn broadcasts(&self, match_all_topics: &[Topic]) -> Result<Vec<Vec<u8>>>;
    fn posted(
        &self,
        match_all_topics: &[Topic],
        dest: [u8; 32],
    ) -> Result<Vec<Vec<u8>>>;
    fn posted_clear(
        &self,
        match_all_topics: &[Topic],
        dest: [u8; 32],
    ) -> Result<Vec<Vec<u8>>>;
    fn broadcasts_stmt(
        &self,
        match_all_topics: &[Topic],
    ) -> Result<Vec<Vec<u8>>>;
    fn posted_stmt(
        &self,
        match_all_topics: &[Topic],
        dest: [u8; 32],
    ) -> Result<Vec<Vec<u8>>>;
    fn posted_clear_stmt(
        &self,
        match_all_topics: &[Topic],
        dest: [u8; 32],
    ) -> Result<Vec<Vec<u8>>>;
    fn submit(
        &self,
        statement: Statement,
        source: StatementSource,
    ) -> SubmitResult;
    fn remove(&self, hash: &Hash) -> Result<()>;
    fn remove_by(&self, who: [u8; 32]) -> Result<()>;

}

Statement store API.

Required Methods

fn statements(&self) -> Result<Vec<(Hash, Statement)>>

Return all statements.

fn take_recent_statements(&self) -> Result<Vec<(Hash, Statement)>>

Return recent statements and clear the internal index.

This consumes and clears the recently received statements, allowing new statements to be collected from this point forward.

fn statement(&self, hash: &Hash) -> Result<Option<Statement>>

Get statement by hash.

fn has_statement(&self, hash: &Hash) -> bool

Check if statement exists in the store

Fast index check without accessing the DB.

fn statement_hashes(&self) -> Vec<Hash>

Return all statement hashes.

fn statements_by_hashes( &self, hashes: &[Hash], filter: &mut dyn FnMut(&Hash, &[u8], &Statement) -> FilterDecision, ) -> Result<(Vec<(Hash, Statement)>, usize)>

Fetch statements by their hashes with a filter callback.

The callback receives (hash, encoded_bytes, decoded_statement) and returns:

Skip: ignore this statement, continue to next

Take: include this statement in the result, continue to next

Abort: stop iteration, return collected statements so far

Returns (statements, number_of_hashes_processed).

fn broadcasts(&self, match_all_topics: &[Topic]) -> Result<Vec<Vec<u8>>>

Return the data of all known statements which include all topics and have no DecryptionKey field.

fn posted( &self, match_all_topics: &[Topic], dest: [u8; 32], ) -> Result<Vec<Vec<u8>>>

Return the data of all known statements whose decryption key is identified as dest (this will generally be the public key or a hash thereof for symmetric ciphers, or a hash of the private key for symmetric ciphers).

fn posted_clear( &self, match_all_topics: &[Topic], dest: [u8; 32], ) -> Result<Vec<Vec<u8>>>

Return the decrypted data of all known statements whose decryption key is identified as dest. The key must be available to the client.

fn broadcasts_stmt(&self, match_all_topics: &[Topic]) -> Result<Vec<Vec<u8>>>

Return all known statements which include all topics and have no DecryptionKey field.

fn posted_stmt( &self, match_all_topics: &[Topic], dest: [u8; 32], ) -> Result<Vec<Vec<u8>>>

Return all known statements whose decryption key is identified as dest (this will generally be the public key or a hash thereof for symmetric ciphers, or a hash of the private key for symmetric ciphers).

fn posted_clear_stmt( &self, match_all_topics: &[Topic], dest: [u8; 32], ) -> Result<Vec<Vec<u8>>>

Return the statement and the decrypted data of all known statements whose decryption key is identified as dest. The key must be available to the client.

The result is for each statement: the SCALE-encoded statement concatenated to the decrypted data.

fn submit(&self, statement: Statement, source: StatementSource) -> SubmitResult

Submit a statement.

fn remove(&self, hash: &Hash) -> Result<()>

Remove a statement from the store.

fn remove_by(&self, who: [u8; 32]) -> Result<()>

Remove all statements authored by who.